pyproject.toml
[tool.poetry]
name = "bot-system"
version = "0.1.0"
description = "Bot System for Liberty Pace Agency"
authors = ["franckess <franckess48@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
appdirs = "1.4.4"
attrs = "23.2.0"
beautifulsoup4 = "4.12.3"
bs4 = "0.0.2"
certifi = "2024.6.2"
charset-normalizer = "3.3.2"
exceptiongroup = "1.2.1"
faker = "26.0.0"
h11 = "0.14.0"
idna = "3.7"
importlib-metadata = "8.0.0"
lxml = "5.2.2"
outcome = "1.3.0.post0"
packaging = "24.1"
pyee = "11.1.0"
pyppeteer = "2.0.0"
pysocks = "1.7.1"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pytz = "2024.1"
requests = "2.32.3"
selenium = "4.22.0"
six = "1.16.0"
sniffio = "1.3.1"
sortedcontainers = "2.4.0"
soupsieve = "2.5"
tqdm = "4.66.4"
trio = "0.25.1"
trio-websocket = "0.11.1"
typing-extensions = "4.12.2"
urllib3 = "1.26.19"
webdriver-manager = "4.0.1"
websocket-client = "1.8.0"
websockets = "10.4"
wsproto = "1.2.0"
zipp = "3.19.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
